Fix skipped message to display properly

+++ lib/common/functions.sh
@@ -1436,7 +1436,9 @@ EOF
             echo "Skipped"
             ;;
         *)
-            if [[ $recreateCA == yes || $recreateKeys == yes || ! -f $etcconf ]]; then
+            if [[ $recreateCA != yes && $recreateKeys != yes && -f $etcconf ]]; then
+                echo "Skipped"
+            else
                 if [[ $httpproto == https ]]; then
                     echo "<VirtualHost *:80>" > "$etcconf"
                     echo "    <FilesMatch \"\.php\$\">" >> "$etcconf"
@@ -1497,8 +1499,6 @@ EOF
                     a2enmod ssl >>$workingdir/error_logs/fog_error_${version}.log 2>&1
                     a2ensite "001-fog" >>$workingdir/error_logs/fog_error_${version}.log 2>&1
                 fi
-            else
-                echo "Skipped"
             fi
             ;;
     esac
@@ -2044,17 +2044,21 @@ downloadfiles() {
     echo "Done"
 }
 configureDHCP() {
+    case $linuxReleaseName in
+        *[Dd][Ee][Bb][Ii][Aa][Nn]*)
+            if [[ $bldhcp -eq 1 ]]; then
+                dots "Setting up and starting DHCP Server (incl. debian 9 fix)"
+                sed -i.fog "s/INTERFACESv4=\"\"/INTERFACESv4=\"$interface\"/g" /etc/default/isc-dhcp-server
+            else
+                dots "Setting up and starting DHCP Server"
+            fi
+            ;;
+        *)
+            dots "Setting up and starting DHCP Server"
+            ;;
+    esac
     case $bldhcp in
         1)
-            case $linuxReleaseName in
-                *[Dd][Ee][Bb][Ii][Aa][Nn]*)
-                    dots "Setting up and starting DHCP Server (incl. debian 9 fix)"
-                    sed -i.fog "s/INTERFACESv4=\"\"/INTERFACESv4=\"$interface\"/g" /etc/default/isc-dhcp-server
-                    ;;
-                *)
-                    dots "Setting up and starting DHCP Server"
-                    ;;
-            esac
             [[ -f $dhcpconfig ]] && cp -f $dhcpconfig ${dhcpconfig}.fogbackup
             serverip=$(ip -4 -o addr show $interface | awk -F'([ /])+' '/global/ {print $4}')
             [[ -z $serverip ]] && serverip=$(/sbin/ifconfig $interface | grep -oE 'inet[:]? addr[:]?([0-9]{1,3}\.){3}[0-9]{1,3}' | awk -F'(inet[:]? ?addr[:]?)' '{print $2}')
